713 moved Property2 into Property
